requirements.txt
llama_index
llmsherpa
streamlit
python-dotenv==1.0.1